changes to examples and cleanup readme demo
