+++ examples/lite_workflow_inference.py
@@ -4,7 +4,7 @@ from minference.lite.inference import InferenceOrchestrator, RequestLimits
 from minference.lite.models import ChatThread, LLMConfig, CallableTool, LLMClient, ResponseFormat, SystemPrompt, StructuredTool
 from typing import List
 from pydantic import BaseModel
-from minference.enregistry import EntityRegistry
+from minference.entity import EntityRegistry
 from minference.caregistry import CallableRegistry
 import statistics
 from minference.utils import msg_dict_to_oai, msg_dict_to_anthropic, parse_json_string
